fix(fit): IMP Copilot review followups
- R/fit.R: fold IMP aliases (`importance_sampling`,
  `importance-sampling`, case-insensitive) to `"imp"` before
  `match.arg`. Partial-prefix matching in `match.arg` can't see
  `importance_sampling` as a prefix of `imp`, so the documented
  aliases used to error at runtime.
- R/fit.R: reorder `tolower`/`gsub` — uppercase input like `"IMP"`
  used to get smashed to `"___"` because the `[^a-z0-9]` strip
  ran before tolower.
- R/fit.R: surface engine-side `imp` chain validation R-side too —
  reject `method = "imp"`, non-terminal `imp`, and duplicated `imp`
  with messages anchored to the R argument, before the engine
  round-trip. Engine guards (covered in ferx-core #66's integration
  tests) remain as a safety net.
- src/rust/src/lib.rs: cast `n_samples: usize` as `f64` (not `i32`)
  when building the R list so a user-set `is_samples > i32::MAX`
  can't silently wrap to a negative number.
- tests/testthat/test-fit.R: fix the inlined `normalize()` helper
  to mirror the new alias-fold/tolower-order logic, add coverage
  for the documented aliases, and add a new test exercising each
  of the four R-side rejection paths via `expect_error`.
- man/ferx_fit.Rd: regenerated via roxygen2 to reflect the alias
  mention in `@param method`.

+++ src/rust/src/lib.rs
@@ -1061,7 +1061,10 @@ fn fit_result_to_list(
             list!(
                 minus2_log_likelihood = is.minus2_log_likelihood,
                 mc_standard_error = is.mc_standard_error,
-                n_samples = is.n_samples as i32,
+                // n_samples is usize in ferx-core; cast to f64 (rather than
+                // i32) so a user-set `is_samples > i32::MAX` can't silently
+                // wrap to a negative number when bridged to R.
+                n_samples = is.n_samples as f64,
                 proposal_df = is.proposal_df,
                 ess_min = is.ess_min,
                 ess_median = is.ess_median,
